Seperated Story contract editors
There is now a contract view and a contract editor window

The Contract view allows you to see and edit the existing contracts. There
is a delete button but it doesn't work yet.

The Contract editor works just like it did before I started playing
with viewing existing contracts.

Not sure whether to keep the "new" section of the contracts but it will
remain there for the time being. The Contract view allows you to create a
new contract.